separate model-handling into a shared part and a main-cli only part, cli-utils no longer depend on counterexamples
