+++ src/storm-cli-utilities/CMakeLists.txt
@@ -11,7 +11,7 @@ target_sources(storm-cli-utilities
 		FILE_SET fs_storm_cli_utilities_headers TYPE HEADERS BASE_DIRS "${PROJECT_SOURCE_DIR}/src" FILES ${STORM_CLI_UTIL_HEADERS})
 target_precompile_headers(storm-cli-utilities REUSE_FROM storm)
 set_target_properties(storm-cli-utilities PROPERTIES VERSION ${STORM_VERSION} SOVERSION ${STORM_VERSION})
-target_link_libraries(storm-cli-utilities PUBLIC storm storm-counterexamples storm-gamebased-ar storm-parsers storm-version-info)
+target_link_libraries(storm-cli-utilities PUBLIC storm storm-parsers storm-version-info)
 set_target_properties(storm-cli-utilities PROPERTIES DEFINE_SYMBOL "") # to avoid problems with pch on linux.
 list(APPEND STORM_TARGETS storm-cli-utilities)
 set(STORM_TARGETS ${STORM_TARGETS} PARENT_SCOPE)
