+++ src/storm-cli/storm-cli.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "storm-cli-utilities/cli.h"
-#include "storm-cli-utilities/model-handling.h"
+#include "storm-cli/model-handling-main-cli.h"
 #include "storm/adapters/IntervalAdapter.h"
 #include "storm/adapters/RationalFunctionAdapter.h"
 #include "storm/adapters/RationalNumberAdapter.h"
