+++ src/storm-counterexamples/counterexamples/MILPMinimalLabelSetGenerator.h
@@ -4,7 +4,7 @@
 
 #include "storm-counterexamples/counterexamples/GuaranteedLabelSet.h"
 #include "storm-counterexamples/counterexamples/HighLevelCounterexample.h"
-#include "storm-counterexamples/settings/modules/CounterexampleGeneratorSettings.h"
+#include "storm/settings/modules/CounterexampleGeneratorSettings.h"
 
 #include "storm/exceptions/InvalidArgumentException.h"
 #include "storm/exceptions/InvalidPropertyException.h"
