+++ src/storm-pomdp-cli/storm-pomdp.cpp
@@ -22,7 +22,9 @@
 #include "storm-pomdp/transformer/PomdpMemoryUnfolder.h"
 #include "storm/analysis/GraphConditions.h"
 #include "storm/api/storm.h"
+#include "storm/exceptions/InvalidPropertyException.h"
 #include "storm/exceptions/NotSupportedException.h"
+#include "storm/exceptions/WrongFormatException.h"
 #include "storm/exceptions/UnexpectedException.h"
 #include "storm/modelchecker/results/ExplicitQualitativeCheckResult.h"
 #include "storm/settings/modules/DebugSettings.h"
