+++ src/storm/settings/modules/CounterexampleGeneratorSettings.cpp
@@ -1,4 +1,4 @@
-#include "storm-counterexamples/settings/modules/CounterexampleGeneratorSettings.h"
+#include "storm/settings/modules/CounterexampleGeneratorSettings.h"
 
 #include "storm/exceptions/InvalidSettingsException.h"
 #include "storm/settings/Argument.h"
